requirements.txt
Django==1.8
fab==1.4.2
ipdb==0.8
djangorestframework==3.1.1
django-filter==0.9.2
